Add Processor-and-assembler as regular folder

+++ stack/stack.h
@@ -2,7 +2,7 @@
 #define STACK_H
 #include <stdlib.h>
 
-typedef size_t stack_elem_t; 
+typedef int stack_elem_t; 
 
 const size_t FRONT_CANARY_STACK = 0x4255463132333445;
 const size_t TAIL_CANARY_STACK = 0x4E45534255464145;
